added new file

+++ 2_VariableaAndDatatype/type_conversion.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int x = 5;
-    int y = 2;
+    float x = 5;
+    float y = 2;
     int sum = 5 / 2;
 
     printf("%d", sum);
